move timeToStr to date util
